+++ Display.cpp
@@ -36,7 +36,7 @@ void Display::DrawHp(VECTOR hpPos, int hitPoint, int maxHp) {
 
 	SetFontSize(20);
 	DrawFormatString((int)hpPos.x - 30, 0, GetColor(255, 0, 255), "HP:");
-	int hpRate = ((float)hitPoint / maxHp) * 100;
+	int hpRate = (int)(((float)hitPoint / maxHp) * 100);
 	for (int i = 0; i < hpRate; i++) {
 		DrawGraph((int)hpPos.x + 4 * i, (int)hpPos.y, hpGraphic, TRUE);
 	}
